+++ src/models/fwcloud/FwCloud.ts
@@ -40,7 +40,6 @@ const fwcError = require('../../utils/error_table');
 
 const tableName: string = 'fwcloud';
 
-
 @Entity(tableName)
 export class FwCloud extends Model {
 
@@ -123,6 +122,30 @@ export class FwCloud extends Model {
 		fs.mkdirpSync(this.getSnapshotDirectoryPath());
 	}
 
+	/**
+	 * Removes all fwcloud trees
+	 *
+	 */
+	public async removeTrees(queryRunner: QueryRunner): Promise<void> {
+		const sqls : string[] = [];
+		
+		// Root nodes.
+		let nodes = await queryRunner.query(`select id from fwc_tree where fwcloud=${this.id} and id_parent is null`);
+
+		// Next levels nodes.
+		while(nodes.length > 0){
+			sqls.unshift(`delete from fwc_tree where id in (${nodes.map(obj => obj.id)})`);
+			nodes = await queryRunner.query(`select id from fwc_tree where id_parent in (${nodes.map(obj => obj.id)})`);
+		}
+
+		// Run queries removing node trees from down up to root nodes.
+		for(let i=0; i < sqls.length; i++)
+			await queryRunner.query(sqls[i]);
+
+		return;
+	}
+		
+	
 	public async remove(options?: RemoveOptions): Promise<this> {
 		const databaseService = await app().getService<DatabaseService>(DatabaseService.name);
 		const queryRunner: QueryRunner = databaseService.connection.createQueryRunner();
@@ -130,9 +153,13 @@ export class FwCloud extends Model {
 		try {
 			await queryRunner.startTransaction();
 
-      // WARNING: Don't use 'SET FOREIGN_KEY_CHECKS=0' and 'SET FOREIGN_KEY_CHECKS=1'
+			// WARNING: Don't use 'SET FOREIGN_KEY_CHECKS=0' and 'SET FOREIGN_KEY_CHECKS=1'
       // This way we make sure that the delete procedure follows the referential integrity of the data base and
       // avoid left rows without correct relations in a table.
+
+			// First remove the Firewall, Objects, Services and CA trees.
+			await this.removeTrees(queryRunner);
+
 			let query =
         // First remove the relational tables for policy rules and the policy rules themselves.
 			 `delete PI from policy_r__interface PI inner join policy_r RULE on RULE.id=PI.rule inner join firewall FW on FW.id=RULE.firewall where FW.fwcloud=${this.id};
@@ -168,11 +195,6 @@ export class FwCloud extends Model {
 
         // Firewall interfaces.
       +`delete I from interface I inner join firewall FW on FW.id=I.firewall where FW.fwcloud=${this.id};`
-
-      // Trees. 
-			+`SET FOREIGN_KEY_CHECKS=0;
-        delete from fwc_tree where fwcloud=${this.id};
-        SET FOREIGN_KEY_CHECKS=1;`
 
         // Clusters and firewalls.
 		  +`delete from firewall where fwcloud=${this.id};
@@ -572,87 +594,6 @@ export class FwCloud extends Model {
 						resolve({ "result": false });
 					}
 				});
-			});
-		});
-	}
-
-	/**
-	 * DELETE Fwcloud
-	 *
-	 * @method deleteFwcloud
-	 *
-	 * @param iduser {Integer}  User identifier
-	 * @param id {Integer}  Fwcloud identifier
-	 * @param {Function} callback    Function callback response
-	 *
-	 *       callback(error, Rows)
-	 *
-	 * @return {CALLBACK RESPONSE}
-	 *
-	 * @example
-	 * #### RESPONSE OK:
-	 *
-	 *       callback(null, {"result": true, "msg": "deleted"});
-	 *
-	 * #### RESPONSE ERROR:
-	 *
-	 *       callback(null, {"result": false});
-	 *
-	 */
-	public static deleteFwcloud(req) {
-		return new Promise((resolve, reject) => {
-			let sql = `SELECT C.* FROM ${tableName} C
-				INNER JOIN user__fwcloud U ON C.id=U.fwcloud
-				WHERE U.user=${req.session.user_id} AND C.id=${req.body.fwcloud}`;
-			req.dbCon.query(sql, async(error, row) => {
-				if (error) return reject(error);
-
-				//If exists Id from fwcloud to remove
-				if (row && row.length > 0) {
-					try {
-						//DELETE ALL OBJECTS FROM CLOUD
-						await this.EmptyFwcloudStandard(req.body.fwcloud);
-						const admins: any = await User.getAllAdminUserIds(req);
-						for(let admin of admins) {
-							await User.disableFwcloudAccess(req.dbCon,admin.id,req.body.fwcloud);
-						}
-					} catch (error) { return reject(error) }
-
-					req.dbCon.query(`DELETE FROM ${tableName} WHERE id=${req.body.fwcloud}`, (error, result) => {
-						if (error) return reject(error);
-						resolve();
-					});
-				} else reject(fwcError.NOT_FOUND);
-			});
-		});
-	}
-
-	private static EmptyFwcloudStandard(fwcloud) {
-		return new Promise((resolve, reject) => {
-			var sqlcloud = "  is null";
-			if (fwcloud !== null)
-				sqlcloud = "= " + fwcloud;
-			db.get(async (error, connection) => {
-				if (error)
-					reject(error);
-
-				try {
-					const databaseService = await app().getService<DatabaseService>(DatabaseService.name);
-					await databaseService.connection.transaction(async transactionalManager => {
-						await transactionalManager.query("SET FOREIGN_KEY_CHECKS = 0");
-						await transactionalManager.query("DELETE I.* from interface I inner join interface__ipobj II on II.interface=I.id inner join ipobj G On G.id=II.ipobj where G.fwcloud" + sqlcloud);
-						await transactionalManager.query("DELETE II.* from interface__ipobj II inner join ipobj G On G.id=II.ipobj where G.fwcloud" + sqlcloud);
-						await transactionalManager.query("DELETE II.* from ipobj__ipobjg II inner join ipobj G On G.id=II.ipobj where G.fwcloud" + sqlcloud);
-						await transactionalManager.query("DELETE FROM ipobj_g where fwcloud" + sqlcloud);
-						await transactionalManager.query("DELETE FROM ipobj where fwcloud" + sqlcloud);
-						await transactionalManager.query("DELETE FROM ipobj where fwcloud" + sqlcloud);
-						await transactionalManager.query("DELETE FROM fwc_tree where fwcloud" + sqlcloud);
-						await transactionalManager.query("SET FOREIGN_KEY_CHECKS = 1");
-					});
-					resolve({ "result": true });
-				} catch (e) {
-					reject(e);
-				}
 			});
 		});
 	}